+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 0.2.2
 =====
 
+- Accept a path for the `:atomic` option to
+  {HandBrake::CLI#output}. If specified, the temporary file will be
+  written to this path rather than the ultimate target directory.
 - Ensure that trace mode prints updating lines (e.g., the encode
   status) that do not end in newlines in a timely fashion.